Split theme into files
