+++ front/src/modules/ui/layout/styles/colors/border.ts
@@ -1,0 +1,11 @@
+export const borderColorsLight = {
+  primaryBorder: 'rgba(0, 0, 0, 0.08)',
+  lightBorder: 'rgba(245, 245, 245, 1)',
+  mediumBorder: '#ebebeb',
+};
+
+export const borderColorsDark = {
+  primaryBorder: 'rgba(255, 255, 255, 0.08)',
+  lightBorder: '#222222',
+  mediumBorder: '#141414',
+};
