+++ front/src/modules/ui/layout/styles/colors/modal.ts
@@ -1,0 +1,3 @@
+export const modalColorsLight = {};
+
+export const modalColorsDark = {};
