+++ front/src/modules/ui/layout/styles/colors/palette.ts
@@ -1,0 +1,1 @@
+export const paletteColors = {};
