+++ front/src/modules/ui/layout/styles/colors/text.ts
@@ -1,0 +1,19 @@
+export const textColorsLight = {
+  text100: '#000',
+  text80: '#333333',
+  text60: '#666',
+  text40: '#999999',
+  text30: '#b3b3b3',
+  text20: '#cccccc',
+  text0: '#fff',
+};
+
+export const textColorsDark = {
+  text100: '#ffffff',
+  text80: '#cccccc',
+  text60: '#999',
+  text40: '#666',
+  text30: '#4c4c4c',
+  text20: '#333333',
+  text0: '#000',
+};
