+++ front/src/modules/ui/layout/styles/colors/transparent.ts
@@ -1,0 +1,17 @@
+export const transparentColorsLight = {
+  primaryBackgroundTransparent: 'rgba(255, 255, 255, 0.8)',
+  secondaryBackgroundTransparent: 'rgba(252, 252, 252, 0.8)',
+  strongBackgroundTransparent: 'rgba(0, 0, 0, 0.16)',
+  mediumBackgroundTransparent: 'rgba(0, 0, 0, 0.08)',
+  lightBackgroundTransparent: 'rgba(0, 0, 0, 0.04)',
+  lighterBackgroundTransparent: 'rgba(0, 0, 0, 0.02)',
+};
+
+export const transparentColorsDark = {
+  primaryBackgroundTransparent: 'rgba(20, 20, 20, 0.8)',
+  secondaryBackgroundTransparent: 'rgba(23, 23, 23, 0.8)',
+  strongBackgroundTransparent: 'rgba(255, 255, 255, 0.09)',
+  mediumBackgroundTransparent: 'rgba(255, 255, 255, 0.06)',
+  lightBackgroundTransparent: 'rgba(255, 255, 255, 0.03)',
+  lighterBackgroundTransparent: 'rgba(255, 255, 255, 0.02)',
+};
